Refactor: move mutations to separate file and rearrange imports on all components
