+++ src/components/Footer/Footer.js
@@ -1,4 +1,3 @@
-import React from 'react'
 import { Link } from "react-router-dom";
 import './Footer.css'
