+++ src/components/Map/Map.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import GoogleMapReact from "google-map-react";
 import LocationPin from "../LocationPin/LocationPin";
 import "./Map.css";
