+++ src/components/UserSavedTripCard/UserSavedTripCard.js
@@ -1,5 +1,5 @@
 import { useMutation } from '@apollo/client';
-import { DELETE_USER_TRIP_MUTATION } from "../../queries/queries";
+import { DELETE_USER_TRIP_MUTATION } from "../../queries/mutations";
 import { USER_TRIPS_QUERY } from "../../queries/queries";
 import { formatDates } from '../../utilities/utilities';
 import "./UserSavedTripCard.css";
