+++ src/queries/queries.js
@@ -20,30 +20,7 @@ const USER_QUERY = gql`
   }
 `;
 
-const CREATE_USER_MUTATION = gql`
-  mutation CreateUser(
-    $name: String!
-    $email: String!
-    $password: String!
-    $passwordConfirmation: String!
-  ) {
-    createUser(
-      input: {
-        name: $name
-        email: $email
-        password: $password
-        passwordConfirmation: $passwordConfirmation
-      }
-    ) {
-      user {
-        id
-        name
-        email
-      }
-      errors
-    }
-  }
-`;
+
 
 const CREATE_SESSION_QUERY = gql`
   query CreateSession($email: String!, $password: String!) {
@@ -88,78 +65,7 @@ const USER_TRIPS_QUERY = gql`
   }
 `;
 
-const CREATE_TRIP_MUTATION = gql`
-  mutation CreateTrip(
-    $userId: ID!
-    $name: String!
-    $destinationName: String!
-    $destinationPlaceId: String!
-    $startDate: String!
-    $startTime: String!
-    $maxAttendees: Int!
-  ) {
-    createTrip(
-      input: {
-        userId: $userId
-        name: $name
-        destinationName: $destinationName
-        destinationPlaceId: $destinationPlaceId
-        startDate: $startDate
-        startTime: $startTime
-        maxAttendees: $maxAttendees
-      }
-    ) {
-      trip {
-        name
-        destinationName
-        destinationPlaceId
-        startTime
-        maxAttendees
-      }
-      errors
-    }
-  }
-`;
 
-const CREATE_USER_TRIP_MUTATION = gql`
-  mutation CreateUserTrip($userId: ID!, $tripId: ID!, $isHost: Boolean!) {
-    createUserTrip(
-      input: { userId: $userId, tripId: $tripId, isHost: $isHost }
-    ) {
-      userTrip {
-        userId
-        tripId
-        user {
-          name
-          email
-        }
-      }
-    }
-  }
-`;
-
-const DELETE_USER_TRIP_MUTATION = gql`
-  mutation DeleteUserTrip($userId: ID!, $tripId: ID!) {
-    deleteUserTrip(
-      input: {
-        userId: $userId
-        tripId: $tripId
-      }
-    ) 
-      {
-        confirm
-      }
-  }
-`;
-
-const DELETE_TRIP_MUTATION = gql`
-  mutation DeleteTrip($id: ID!) {
-    deleteTrip(input: { id: $id }) {
-      confirm
-      errors
-    }
-  }
-`;
 
 const MUSEUMS_QUERY = gql`
   query Museums($city: String!, $state: String!, $zipcode: String!) {
@@ -199,14 +105,9 @@ const MUSEUM_QUERY = gql`
 export {
   USERS_QUERY,
   USER_QUERY,
-  CREATE_USER_MUTATION,
   CREATE_SESSION_QUERY,
   TRIPS_QUERY,
   USER_TRIPS_QUERY,
-  CREATE_TRIP_MUTATION,
-  CREATE_USER_TRIP_MUTATION,
-  DELETE_USER_TRIP_MUTATION,
-  DELETE_TRIP_MUTATION,
   MUSEUMS_QUERY,
   MUSEUM_QUERY,
 };
